pyproject.toml
[project]
name = "rossman"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "joblib>=1.5.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pymongo>=4.12.1",
    "python-dotenv>=1.1.0",
    "scikit-learn==1.5.2",
]
